backend/requirements.txt
fastapi
uvicorn
google-genai
numpy
python-multipart
python-dotenv
httpx
pydantic-settings
